1.1 Auth Middleware & Routing

+++ backend/src/controllers/notes.ts
@@ -2,10 +2,15 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from '../utils/assertIsDefined';
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
+
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authenticatedUserId);
+
+    const notes = await NoteModel.find({userId: authenticatedUserId}).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -14,8 +19,11 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid Note id");
     }
@@ -24,6 +32,10 @@ export const getNote: RequestHandler = async (req, res, next) => {
 
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(403, "Note does not belong to user");
     }
 
     res.status(200).json(note);
@@ -46,13 +58,17 @@ export const createNote: RequestHandler<
   unknown
 > = async (req, res, next) => {
   const { title, text, author, assignedTo } = req.body;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!title) {
       throw createHttpError(400, "Title is required");
     }
 
     const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
       title: title,
       text: text,
       author: author,
@@ -88,8 +104,11 @@ export const updateNote: RequestHandler<
     author: newAuthor,
     assignedTo: newAssignedTo,
   } = req.body;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid Note id");
     }
@@ -101,6 +120,10 @@ export const updateNote: RequestHandler<
 
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(403, "Note does not belong to user");
     }
 
     if (newTitle !== undefined) note.title = newTitle;
@@ -117,8 +140,12 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid Note id");
     }
@@ -127,6 +154,11 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
     if (!note) {
       throw createHttpError(404, "Note not found");
     }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(403, "Note does not belong to user");
+    }
+
 
     await note.deleteOne();
